Move predefined errors into their own directory
