+++ src/error/predefined/StatesBranchFailedError.ts
