+++ src/error/predefined/StatesExceedToleratedFailureThresholdError.ts
