+++ src/error/predefined/StatesHeartbeatTimeoutError.ts
