+++ src/error/predefined/StatesIntrinsicFailureError.ts
