+++ src/error/predefined/StatesItemReaderFailedError.ts
