+++ src/error/predefined/StatesPermissionsError.ts
