+++ src/error/predefined/StatesResultPathMatchFailureError.ts
