+++ src/error/predefined/StatesResultWriterFailedError.ts
