+++ src/error/predefined/StatesTaskFailedError.ts
