+++ src/stateMachine/InputOutputProcessing.ts
@@ -2,7 +2,7 @@ import type { PayloadTemplate } from '../typings/InputOutputProcessing';
 import type { JSONValue } from '../typings/JSONValue';
 import { isPlainObj } from '../util';
 import { jsonPathQuery } from './JsonPath';
-import { StatesResultPathMatchFailureError } from '../error/StatesResultPathMatchFailureError';
+import { StatesResultPathMatchFailureError } from '../error/predefined/StatesResultPathMatchFailureError';
 import cloneDeep from 'lodash/cloneDeep.js';
 import set from 'lodash/set.js';
